fixed problems
