+++ app/src/main/res/layout/activity_main.xml
@@ -1,4 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
+
 <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
@@ -7,13 +8,14 @@
 
     <RelativeLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="wrap_content">
 
         <ImageView
             android:id="@+id/header"
             android:layout_width="match_parent"
             android:layout_height="200dp"
             android:adjustViewBounds="true"
+            android:contentDescription="@string/content_description_for_my_photo"
             android:scaleType="centerCrop"
             tools:src="@drawable/avatarka" />
 
@@ -32,7 +34,8 @@
                 android:drawableLeft="@drawable/ic_person_black_24dp"
                 android:drawableStart="@drawable/ic_person_black_24dp"
                 android:text="@string/example_name"
-                android:textSize="@dimen/text_size" />
+                android:textSize="@dimen/text_size"
+                tools:ignore="RtlHardcoded" />
 
             <View
                 android:layout_width="match_parent"
@@ -46,7 +49,8 @@
                 android:drawableLeft="@drawable/ic_mood_black_24dp"
                 android:drawableStart="@drawable/ic_mood_black_24dp"
                 android:text="@string/example_mood"
-                android:textSize="@dimen/text_size" />
+                android:textSize="@dimen/text_size"
+                tools:ignore="RtlHardcoded" />
 
             <View
                 android:layout_width="match_parent"
@@ -60,7 +64,8 @@
                 android:drawableLeft="@drawable/ic_cake_black_24dp"
                 android:drawableStart="@drawable/ic_cake_black_24dp"
                 android:text="@string/example_birth"
-                android:textSize="@dimen/text_size" />
+                android:textSize="@dimen/text_size"
+                tools:ignore="RtlHardcoded" />
 
             <View
                 android:layout_width="match_parent"
@@ -84,6 +89,7 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:hint="@string/hint_for_message"
+                android:inputType="text"
                 android:textSize="@dimen/text_size" />
 
             <Button
@@ -109,15 +115,20 @@
                 android:id="@+id/vk"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
+                android:layout_marginEnd="@dimen/margin_between_comonents"
                 android:layout_marginRight="@dimen/margin_between_comonents"
-                android:src="@drawable/vk" />
+                android:contentDescription="@string/vk"
+                android:src="@drawable/vk"
+                tools:ignore="RtlHardcoded" />
 
             <ImageView
                 android:id="@+id/twitter"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
+                android:contentDescription="@string/twitter"
                 android:src="@drawable/twitter" />
         </LinearLayout>
 
     </RelativeLayout>
+
 </ScrollView>
