+++ app/src/main/res/layout/activity_news_list.xml
@@ -10,6 +10,6 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:layout_margin="@dimen/half_standart_margin"
-        android:orientation="vertical"></android.support.v7.widget.RecyclerView>
+        android:orientation="vertical" />
 
 </RelativeLayout>
